+++ packages/firebase_dynamic_links/lib/src/firebase_dynamic_links.dart
@@ -30,9 +30,10 @@ class FirebaseDynamicLinks {
   /// This method always returns a Future. That Future completes to null if
   /// there is no pending dynamic link or any call to this method after the
   /// the first attempt.
-  Future<PendingDynamicLinkData> getLaunchLink() async {
-    final Map<String, dynamic> linkData = await channel
-        .invokeMapMethod<String, dynamic>('FirebaseDynamicLinks#getLaunchLink');
+  Future<PendingDynamicLinkData> getInitialLink() async {
+    final Map<String, dynamic> linkData =
+        await channel.invokeMapMethod<String, dynamic>(
+            'FirebaseDynamicLinks#getInitialLink');
     return getPendingDynamicLinkDataFromMap(linkData);
   }